[ir][refactor] Move all Expression subclasses to `frontend_ir.h`
